Inherit from AggregateRoot
