+++ spec/support/test_domain.rb
@@ -38,35 +38,3 @@ end
 RSpec.configure do |config|
   config.include(EventHelpers)
 end
-
-class Category
-  include Eventory::AggregateRoot
-
-  def initialize(id, events = [])
-    @item_added_events = []
-    @item_removed_events = []
-    @added_and_removed_events = []
-    super
-  end
-
-  on ItemAdded do |event|
-    @item_added_events << event
-  end
-
-  on ItemRemoved do |event|
-    @item_removed_events << event
-  end
-
-  on ItemAdded, ItemRemoved do |event|
-    @added_and_removed_events << event
-  end
-
-  def add_item(item)
-    apply_event ItemAdded.new(item_id: item.id, name: item.name)
-  end
-
-  attr_reader :item_added_events,
-              :item_removed_events,
-              :added_and_removed_events
-
-end
